cross_layer_ids/requirements.txt
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
scipy>=1.7.0
pyserial>=3.5
streamlit>=1.28.0
streamlit-autorefresh>=1.0.0
plotly>=5.15.0
matplotlib>=3.5.0
